+++ uMB/VS1053.c
@@ -12,8 +12,7 @@ WORD  writeMP3( BYTE b)
     return getcSPI();
 }
 
-
-void writeMP3Register( BYTE reg, unsigned w)
+void writeMP3Register( BYTE reg, WORD w)
 {
     MP3_DCS_Disable();              // disable data bus
     MP3_CS_Enable();                // enable command bus
